Extract text editor utils
